requirements.txt
python-telegram-bot
tinydb
schedule
ray[serve]